Several fixes for the alpaka implementation of FCS-GPU (#5)
* FastCaloSimAnalyzer/FastCaloGpu/src/CMakeLists.txt
* FastCaloSimAnalyzer/FastCaloGpu/FastCaloGpu/HostDevDef.h

A hack to prevent the exposure of alpaka definitions outside
FastCaloGpu library via HostDevDef.h

* FastCaloSimAnalyzer/FastCaloGpu/FastCaloGpu/AlpakaDefs.h

Fixed the definition of FccDefaultAcc for C++ threads and serial backends

* FastCaloSimAnalyzer/FastCaloGpu/FastCaloGpu/Rand4Hits.h
* FastCaloSimAnalyzer/FastCaloGpu/src/CaloGpuGeneral_al.cxx
* FastCaloSimAnalyzer/FastCaloGpu/src/CaloGpuGeneral_fnc.cxx
* FastCaloSimAnalyzer/FastCaloGpu/src/GeoLoadGpu_al.cxx
* FastCaloSimAnalyzer/FastCaloGpu/src/GeoRegion.cxx

Fixed compiler issues uncovered when building for alpaka C++ threads backend

+++ FastCaloSimAnalyzer/FastCaloGpu/src/CaloGpuGeneral_al.cxx
@@ -42,7 +42,7 @@ namespace CaloGpuGeneral_al {
 				  ) const -> void
     {
       auto const idx = alpaka::getIdx<alpaka::Grid, alpaka::Threads>(acc)[0];
-      if(idx < nhits) {
+      if(idx < (unsigned)nhits) {
 	Hit hit;
 	hit.E() = E;
 	
@@ -180,7 +180,7 @@ namespace CaloGpuGeneral_al {
 #ifdef DUMP_HITCELLS
     std::cout << "hitcells: " << args.ct << "  nhits: " << nhits << "  E: " << E << "\n";
     std::map<unsigned int,float> cm;
-    for (int i=0; i<args.ct; ++i) {
+    for (unsigned i=0; i<args.ct; ++i) {
       cm[args.hitcells_E_h[i].cellid] = args.hitcells_E_h[i].energy;
     }
     for (auto &em: cm) {
